Also fix speculation in checked expression
