+++ src/Compilers/CSharp/Portable/Compilation/MemberSemanticModel.cs
@@ -340,6 +340,10 @@ namespace Microsoft.CodeAnalysis.CSharp
                 {
                     binder = rootBinder.GetBinder(current);
                 }
+                else if (current is CheckedExpressionSyntax)
+                {
+                    binder = rootBinder.GetBinder(current);
+                }
                 else
                 {
                     // If this ever breaks, make sure that all callers of
